feat: 完成薪酬标准模块

+++ src/main/java/com/gdou/pojo/entity/SalaryStandardItem.java
@@ -1,16 +1,17 @@
 package com.gdou.pojo.entity;
 
 import com.baomidou.mybatisplus.annotation.TableName;
+import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * 薪酬标准与薪酬项目关系表
+ *
  * @TableName salary_standard_item
  */
-@TableName(value ="salary_standard_item")
+@TableName(value = "salary_standard_item")
 @Data
 public class SalaryStandardItem implements Serializable {
     /**
@@ -26,5 +27,5 @@ public class SalaryStandardItem implements Serializable {
     /**
      * 该项目对应的金额
      */
-    private BigDecimal money;
+    private Double money;
 }
